membuat kondisi saat delete user

+++ repository/user/user.go
@@ -37,17 +37,22 @@ func (ur *UserRepository) GetUser(id int) (_entities.User, []_entitiesproduct.Pr
 	return users, product, int(tx.RowsAffected), nil
 }
 
-func (ur *UserRepository) DeleteUser(id int) (_entities.User, error) {
+func (ur *UserRepository) DeleteUser(id int) (_entities.User, int, error) {
 	var user _entities.User
+	var product []_entitiesproduct.Product
+	err := ur.database.Where("user_id", id).Find(&product)
+	if len(product) != 0 {
+		return user, 0, err.Error
+	}
 	tx := ur.database.Delete(&user, id)
 	if tx.Error != nil {
-		return user, tx.Error
+		return user, 0, tx.Error
 	}
 	if tx.RowsAffected == 0 {
-		return user, tx.Error
+		return user, 0, tx.Error
 
 	}
-	return user, nil
+	return user, int(tx.RowsAffected), nil
 }
 func (ur *UserRepository) CreateUser(user _entities.User) (_entities.User, int, error) {
 	if user.Username == "" {
